feat(fe): make my api directory
